Speed up non-spatial paster commands.
Because plugins are loaded for every paster command, regardless of which
paster command is being run, when lots of plugins are installed startup
time can be substantial.

To fix this, making sure that plugin.py in various extensions doesn't
import unnecessary (and potentially slow) modules, this PR moves the
imports of some function into the function where they are called.

This reduces paster startup time (for me) by up to .5 seconds, which
while not significant by itself, contributes to my current 8 second
startup time.