Remove package entirely that are marked for delete
If the harvest is managing the datasets in ckan, it seems that the
harvest source should be the "source of truth".
If this is the case, we shouldn't need "revive" capability of soft
removing packages/datasets in ckan. I propose to actually purge
the dataset within ckan.
Since it's difficult/nearly impossible to track these files without a
unique id, sometimes the harvester will delete and create a new item if
the waf or files change in any way. This would keep that behind the
scenes, and allow the end user to get to the same dataset at the old URL