Use leaflet-providers library to handle base tile layer
Rather than manually handle the base Leaflet layer, defer it to the
Leaflet-providers library [1], which allows integrating many different
providers. We no longer default to any tile set provider, as they all
require registration. If no provider is set in config, a notice is
displayed to users pointing to the documentation page.

To set a provider in most cases only two options are needed, eg:

    ckanext.spatial.common_map.type = Stadia.StamenTerrain
    ckanext.spatial.common_map.apikey = <api_key>

or:

    ckanext.spatial.common_map.type = MapBox
    ckanext.spatial.common_map.id = <map_id>
    ckanext.spatial.common_map.accessToken = <access_token>

The old `mapbox`, `wms` and `custom` modes are kept for backwards
compatibility.

[1] https://github.com/leaflet-extras/leaflet-providers